+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Current
 - Ensure `basePath` is always a path
 - Hide Namespaces with all hidden Resources from Swagger documentation
 - Per route Swagger documentation for multiple routes on a ``Resource``
+- Fix Swagger `duplicate mapping key` problem from conflicts between response codes given as string or integer (:issue`661`)
 
 0.12.1 (2018-09-28)
 -------------------